+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 * Add :create option to the #write method to control creating any missing directories
 * Add ability to specify a default value for a missing key in the #delete method
+* Add #env_separator for setting a string to separate parts in environment variable name
 
 ### Changed
 * Change #delete to allow removing any subkey of a deeply nested key